chore: 프로젝트 core 패키지 생성 및 변경
