+++ src/main/java/org/indoles/memberserviceserver/core/controller/interfaces/BuyerOnly.java
@@ -1,4 +1,4 @@
-package org.indoles.memberserviceserver.controller.interfaces;
+package org.indoles.memberserviceserver.core.controller.interfaces;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
