+++ src/main/java/org/indoles/memberserviceserver/core/dto/request/SignUpRequestInfo.java
@@ -1,10 +1,9 @@
-package org.indoles.memberserviceserver.dto.request;
+package org.indoles.memberserviceserver.core.dto.request;
 
-import org.indoles.memberserviceserver.entity.exception.MemberException;
+import org.indoles.memberserviceserver.core.entity.exception.MemberException;
+import org.indoles.memberserviceserver.core.entity.exception.MemberExceptionCode;
 
 import java.util.Objects;
-
-import static org.indoles.memberserviceserver.entity.exception.MemberExceptionCode.*;
 
 public record SignUpRequestInfo(
         String signUpId,
@@ -20,7 +19,7 @@ public record SignUpRequestInfo(
 
     private void validateNotNull(Object value, String fieldName) {
         if (Objects.isNull(value)) {
-            throw new MemberException(FIELD_CANNOT_BE_NULL, fieldName, fieldName);
+            throw new MemberException(MemberExceptionCode.FIELD_CANNOT_BE_NULL, fieldName, fieldName);
         }
     }
 }
