+++ src/test/java/org/indoles/memberserviceserver/entity/MemberEntityTest.java
@@ -1,6 +1,7 @@
 package org.indoles.memberserviceserver.entity;
 
 
+import org.indoles.memberserviceserver.core.entity.MemberEntity;
 import org.indoles.memberserviceserver.fixture.MemberFixture;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -8,7 +9,6 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.indoles.memberserviceserver.entity.enums.Role.*;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 class MemberEntityTest {
